feat: do not create PsiWrapper again.

+++ src/main/kotlin/io/github/xyzboom/ssreducer/kotlin/GroupElements.kt
@@ -3,6 +3,7 @@ package io.github.xyzboom.ssreducer.kotlin
 import com.intellij.openapi.project.Project
 import com.intellij.psi.JavaRecursiveElementVisitor
 import com.intellij.psi.PsiAssignmentExpression
+import com.intellij.psi.PsiCall
 import com.intellij.psi.PsiCallExpression
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiClassType
@@ -44,7 +45,7 @@ class GroupElements(
                 fun enterElement(element: PsiElement) {
                     stack.addFirst(element)
                     if (isDecl(element)) {
-                        elements[PsiWrapper(element, true)] = stack.size
+                        elements[PsiWrapper.of(element)] = stack.size
                     }
                     maxLevel = max(maxLevel, stack.size)
                 }
@@ -97,9 +98,9 @@ class GroupElements(
                     return
                 }
                 val oriElement = PsiTreeUtil.findSameElementInCopy(element, oriFile)
-                val oriLevel = fromElements[PsiWrapper(oriElement)]
+                val oriLevel = fromElements[PsiWrapper.of(oriElement)]
                 if (oriLevel != null) {
-                    copiedElements[PsiWrapper(element)] = oriLevel
+                    copiedElements[PsiWrapper.of(element)] = oriLevel
                 }
                 return
             }
@@ -167,26 +168,25 @@ class GroupElements(
         get() = if (this is PsiClassType) {
             val clazz = resolve()
             if (clazz != null) {
-                val wrapper = PsiWrapper(clazz)
-                if (wrapper in elementsInOriProgram && wrapper !in elements) {
-                    "((Object) null)"
+                if (clazz.shouldBeDeleted()) {
+                    "(Object) null"
                 } else {
-                    "((${clazz.qualifiedName!!}) null)"
+                    "(${clazz.qualifiedName!!}) null"
                 }
-            } else canonicalText
+            } else "(${canonicalText}) null"
         } else if (this is PsiPrimitiveType) when {
-            this === PsiTypes.byteType() -> "(java.lang.Byte.parseByte(\"1\"))"
-            this === PsiTypes.charType() -> "('1')"
-            this === PsiTypes.doubleType() -> "(1.0)"
-            this === PsiTypes.floatType() -> "(1.0f)"
-            this === PsiTypes.intType() -> "(1)"
-            this === PsiTypes.longType() -> "(1L)"
-            this === PsiTypes.shortType() -> "(java.lang.Short.parseShort(\"1\"))"
-            this === PsiTypes.booleanType() -> "(true)"
-            this === PsiTypes.voidType() || this === PsiTypes.nullType() -> "(null)"
+            this === PsiTypes.byteType() -> "java.lang.Byte.parseByte(\"1\")"
+            this === PsiTypes.charType() -> "'1'"
+            this === PsiTypes.doubleType() -> "1.0"
+            this === PsiTypes.floatType() -> "1.0f"
+            this === PsiTypes.intType() -> "1"
+            this === PsiTypes.longType() -> "1L"
+            this === PsiTypes.shortType() -> "java.lang.Short.parseShort(\"1\")"
+            this === PsiTypes.booleanType() -> "true"
+            this === PsiTypes.voidType() || this === PsiTypes.nullType() -> "null"
             else -> throw NoWhenBranchMatchedException()
         }
-        else canonicalText
+        else "(${canonicalText}) null"
 
 
     fun getDefaultValueTextForMethod(method: PsiMethod): String {
@@ -194,26 +194,38 @@ class GroupElements(
         if (returnType != null) {
             return returnType.defaultValue
         } else {
-            val defaultClassName = method.containingClass?.qualifiedName
-            return defaultClassName ?: "((Object) null)"
+            val defaultClassName = method.containingClass?.qualifiedName ?: "Object"
+            return "(${defaultClassName}) null"
         }
     }
 
     fun PsiElement.replaceOrDeleteParentStatement(createNewElement: () -> PsiElement) {
         val parent = parent
         if (parent is PsiExpressionStatement && parent.expression === this) {
-            parent.delete()
+            val newElement = createNewElement()
+            if (newElement.canBeStatement()) {
+                replace(newElement)
+            } else {
+                parent.delete()
+            }
         } else {
             replace(createNewElement())
         }
     }
 
     private fun editElementRef2Method(element: PsiElement, target: PsiMethod) {
-        val parent = element.parent ?: return
-        when (parent) {
-            is PsiCallExpression -> {
-                parent.replaceOrDeleteParentStatement {
-                    javaParserFacade.createExpressionFromText(getDefaultValueTextForMethod(target), parent.context)
+        when {
+            element.parent is PsiCallExpression -> {
+                element.parent.replaceOrDeleteParentStatement {
+                    javaParserFacade.createExpressionFromText(
+                        getDefaultValueTextForMethod(target), element.parent.context
+                    )
+                }
+            }
+
+            element is PsiCall -> {
+                element.replaceOrDeleteParentStatement {
+                    javaParserFacade.createExpressionFromText(getDefaultValueTextForMethod(target), element.context)
                 }
             }
         }
@@ -231,7 +243,7 @@ class GroupElements(
             if (parent.lExpression === element) {
                 val rExpr = parent.rExpression
                 if (rExpr != null) {
-                    parent.replace(rExpr)
+                    parent.replaceOrDeleteParentStatement { rExpr }
                 } else {
                     // the assignment expression is unfinished.
                     // for example: `A a = ;`
@@ -274,17 +286,65 @@ class GroupElements(
         return false
     }
 
+    fun PsiElement?.shouldBeDeleted(): Boolean {
+        val wrapper = PsiWrapper.of(this ?: return false)
+        return wrapper in elementsInOriProgram && wrapper !in elements
+    }
+
     fun reconstructDependencies() {
         @Suppress("UNCHECKED_CAST")
         val files = elements.keys.filter { it.element is PsiFile }.map { it.element } as List<PsiFile>
         // we must resolve reference first. Otherwise, the reference will lose after delete.
         val needEdit = mutableListOf<Pair<PsiElement, PsiElement>>()
+        val needDeleteElements = mutableSetOf<PsiElement>()
+
+        for (file in files) {
+            fun recordNeedDelete(element: PsiElement) {
+                if (!element.isValid) return
+                if (!isDecl(element)) return
+                if (element.shouldBeDeleted()) {
+                    needDeleteElements.add(element)
+                }
+                return
+            }
+
+            val visitor = if (file is PsiJavaFile) {
+                object : JavaRecursiveElementVisitor() {
+                    override fun visitElement(element: PsiElement) {
+                        recordNeedDelete(element)
+                        super.visitElement(element)
+                    }
+                }
+            } else {
+                object : KtVisitorVoid() {
+                    override fun visitElement(element: PsiElement) {
+                        recordNeedDelete(element)
+                        super.visitElement(element)
+                    }
+                }
+            }
+            file.accept(visitor)
+        }
+
         for (file in files) {
             fun recordNeedEdit(element: PsiElement) {
                 val targets = element.references.mapNotNull { it.resolve() ?: return@mapNotNull null }
-                for (target in targets) {
-                    val wrapTarget = PsiWrapper(target)
-                    if (wrapTarget in elementsInOriProgram && wrapTarget !in elements) {
+                val callTarget = if (element is PsiCall) {
+                    val method = element.resolveMethod()
+                    if (method != null && !method.shouldBeDeleted()) {
+                        @Suppress("UnstableApiUsage")
+                        for ((i, param) in method.parameters.withIndex()) {
+                            val sourceParam = param.sourceElement
+                            if (sourceParam.shouldBeDeleted()) {
+                                element.argumentList?.expressions[i]?.let { needDeleteElements.add(it) }
+                            }
+                        }
+                    }
+                    method
+                } else null
+                val allTargets = if (callTarget != null) targets + callTarget else targets
+                for (target in allTargets) {
+                    if (target.shouldBeDeleted()) {
                         if (shouldEdit(element, target)) {
                             needEdit.add(element to target)
                             break
@@ -297,11 +357,14 @@ class GroupElements(
                 object : JavaRecursiveElementVisitor() {
                     override fun visitElement(element: PsiElement) {
                         if (!element.isValid) return
+                        if (element in needDeleteElements) return
                         recordNeedEdit(element)
                         super.visitElement(element)
                     }
 
                     override fun visitAssignmentExpression(expression: PsiAssignmentExpression) {
+                        if (!expression.isValid) return
+                        if (expression in needDeleteElements) return
                         // we need to edit right first because sometimes
                         // we may replace the whole assignment with the right
                         expression.rExpression?.accept(this)
@@ -313,34 +376,6 @@ class GroupElements(
                     override fun visitElement(element: PsiElement) {
                         if (!element.isValid) return
                         recordNeedEdit(element)
-                        super.visitElement(element)
-                    }
-                }
-            }
-            file.accept(visitor)
-        }
-
-        val needDeleteElements = mutableSetOf<PsiElement>()
-        for (file in files) {
-            fun recordNeedDelete(element: PsiElement) {
-                if (!element.isValid) return
-                if (!isDecl(element)) return
-                if (PsiWrapper(element) !in elements) {
-                    needDeleteElements.add(element)
-                }
-                return
-            }
-            val visitor = if (file is PsiJavaFile) {
-                object : JavaRecursiveElementVisitor() {
-                    override fun visitElement(element: PsiElement) {
-                        recordNeedDelete(element)
-                        super.visitElement(element)
-                    }
-                }
-            } else {
-                object : KtVisitorVoid() {
-                    override fun visitElement(element: PsiElement) {
-                        recordNeedDelete(element)
                         super.visitElement(element)
                     }
                 }
